+++ src/types.ts
@@ -1,7 +1,7 @@
-export type NSSubCallback<T = any> = (data: T) => void
-export type AnySubCallback<NS = string, T = any> = (namespace: NS, data: T) => void
-export type NSSubs<T> = Set<NSSubCallback<T>>
-export type NSSubsMap<NS = string, T = any> = Map<NS, NSSubs<T>>
-export type AnySubsSet<NS = string, T = any> = Set<AnySubCallback<NS, T>>
-export type Middleware<NS = string, T= any> = (namespace: NS, data?: T) => T
-export type MiddlewareSet<NS = string, T = any> = Set<Middleware<NS, T>>
+export type NSSubCallback<K extends keyof DT, DT> = (data: DT[K]) => void
+export type AnySubCallback<DT> = <K extends keyof DT>(namespace: K, data?: DT[K]) => void
+export type NSSubs<DT> = Set<NSSubCallback<keyof DT, DT>>
+export type NSSubsMap<DT> = Map<keyof DT, NSSubs<DT>>
+export type AnySubsSet<DT> = Set<AnySubCallback<DT>>
+export type Middleware<DT> = <K extends keyof DT>(namespace: K, data: DT[K]) => DT[K]
+export type MiddlewareSet<DT> = Set<Middleware<DT>>